tambah notif

+++ views/daftar_dokter.php
@@ -171,7 +171,7 @@ include '../includes/db.php';
                             <td>{$row['phone']}</td>
                             <td class='action-buttons'>
                                 <a href='edit_dokter.php?id={$row['id']}' class='btn btn-warning'>Edit</a>
-                                <a href='../modules/hapus_dokter.php?id={$row['id']}' class='btn btn-danger'>Hapus</a>
+                                <a href='../modules/hapus_dokter.php?id={$row['id']}' class='btn btn-danger' onclick='return confirmDelete()'>Hapus</a>
                             </td>
                         </tr>";
                 }
@@ -225,6 +225,11 @@ include '../includes/db.php';
             content.style.marginLeft = "250px";
         }
     }
+
+    function confirmDelete() {
+    var confirmation = confirm("Apakah Anda yakin ingin menghapus data?");
+    return confirmation;
+}
 </script>
 </body>
 </html>
