+++ views/daftar_inventaris.php
@@ -178,7 +178,7 @@ include '../includes/db.php';
                         <td>{$row['expiration_date']}</td>
                         <td>
                             <a href='edit_inventaris.php?id={$row['id']}' class='btn btn-warning'>Edit</a>
-                            <a href='../modules/hapus_inventaris.php?id={$row['id']}' class='btn btn-danger'>Hapus</a>
+                            <a href='../modules/hapus_inventaris.php?id={$row['id']}' class='btn btn-danger' onclick='return confirmDelete()'>Hapus</a>
                         </td>
                     </tr>";
             }
@@ -232,6 +232,12 @@ include '../includes/db.php';
             content.style.marginLeft = "250px";
         }
     }
+    
+    function confirmDelete() {
+    var confirmation = confirm("Apakah Anda yakin ingin menghapus data?");
+    return confirmation;
+}
 </script>
+
 </body>
 </html>
